requirements.txt
coverage
django-nose
mock
mock-django
webtest
pillow
django_facebook
google
gcloud


